feat: improve and refactor code background code structure
