+++ source/scripts/Background/events/InactivityTimer.ts
@@ -1,19 +1,10 @@
-﻿export const startInactivityTimer = (autolockMinutes: number) => {
+﻿export const startInactivityTimer = async (autolockMinutes: number) => {
   const autoLockMilliseconds = autolockMinutes * 60 * 1000;
-  let lastActivityTimestamp = Date.now();
+
+  await chrome.storage.local.set({ autoLockMilliseconds });
 
   const resetTimer = () => {
-    lastActivityTimestamp = Date.now();
-  };
-
-  const checkInactivity = () => {
-    const currentTime = Date.now();
-    if (currentTime - lastActivityTimestamp >= autoLockMilliseconds) {
-      chrome.runtime.sendMessage({
-        type: 'lock_wallet',
-        target: 'background',
-      });
-    }
+    chrome.storage.local.set({ lastActivityTimestamp: Date.now() });
   };
 
   // DOM Events
@@ -22,6 +13,22 @@
   events.forEach((event) => {
     document.addEventListener(event, resetTimer);
   });
+  await chrome.alarms.clear('check_inactivity');
+  await chrome.alarms.create('check_inactivity', { periodInMinutes: 0.5 });
+};
 
-  setInterval(checkInactivity, 1000);
+export const checkInactivity = async () => {
+  const { lastActivityTimestamp, autoLockMilliseconds } =
+    await chrome.storage.local.get([
+      'lastActivityTimestamp',
+      'autoLockMilliseconds',
+    ]);
+
+  const currentTime = Date.now();
+  if (currentTime - lastActivityTimestamp >= autoLockMilliseconds) {
+    await chrome.runtime.sendMessage({
+      type: 'lock_wallet',
+      target: 'background',
+    });
+  }
 };
