--- source/scripts/Background/events/onConnect.ts
@@ -1,1 +1,0 @@
-﻿export {};
