--- source/scripts/Background/events/onInstalled.ts
@@ -1,1 +1,0 @@
-﻿export {};
