--- source/scripts/Background/events/onMessage.ts
@@ -1,1 +1,0 @@
-﻿export {};
