--- source/scripts/Background/events/onPaliRequests.ts
@@ -1,1 +1,0 @@
-﻿export {};
