--- source/scripts/Background/events/onPendingTransactionsPolling.ts
@@ -1,1 +1,0 @@
-﻿export {};
