--- source/scripts/Background/events/onPolling.ts
@@ -1,1 +1,0 @@
-﻿export {};
