--- source/scripts/Background/events/onStartup.ts
@@ -1,1 +1,0 @@
-﻿export {};
