+++ source/scripts/Background/handlers/handlePendingTransactionsPolling.ts
@@ -1,0 +1,3 @@
+export const handlePendingTransactionsPolling = () => {
+  chrome.runtime.sendMessage({ type: 'startPendingTransactionsPolling' });
+};
