+++ source/scripts/Background/handlers/handleStartPolling.ts
@@ -1,0 +1,3 @@
+export const handleStartPolling = () => {
+  chrome.runtime.sendMessage({ type: 'startPolling' });
+};
