+++ source/scripts/Background/utils/checkForPendingTransactions.ts
@@ -1,0 +1,50 @@
+import { getController } from 'scripts/Background';
+import { IEvmTransactionResponse } from 'scripts/Background/controllers/transactions/types';
+import { isPollingRunNotValid } from 'scripts/Background/utils/isPollingRunNotValid';
+import store from 'state/store';
+import { TransactionsType } from 'state/vault/types';
+
+export async function checkForPendingTransactionsUpdate() {
+  const { accounts, activeAccount, activeNetwork, isBitcoinBased } =
+    store.getState().vault;
+
+  if (isPollingRunNotValid() || isBitcoinBased) {
+    return;
+  }
+
+  const currentAccountTransactions =
+    (accounts[activeAccount.type]?.[activeAccount.id]?.transactions[
+      TransactionsType.Ethereum
+    ]?.[activeNetwork.chainId] as IEvmTransactionResponse[]) ?? [];
+
+  const pendingTransactions = currentAccountTransactions?.filter(
+    (transaction) => transaction.confirmations === 0
+  );
+
+  if (pendingTransactions.length === 0) {
+    return;
+  }
+
+  const maxTransactionsToSend = 3;
+
+  const cooldownTimeMs = 60 * 1000; //1 minute
+
+  for (let i = 0; i < pendingTransactions.length; i += maxTransactionsToSend) {
+    const batchTransactions = pendingTransactions.slice(
+      i,
+      i + maxTransactionsToSend
+    );
+
+    const { wallet } = getController();
+
+    wallet.validatePendingEvmTransactions({
+      activeNetwork,
+      activeAccount,
+      pendingTransactions: batchTransactions,
+    });
+
+    if (i + maxTransactionsToSend < pendingTransactions.length) {
+      await new Promise((resolve) => setTimeout(resolve, cooldownTimeMs));
+    }
+  }
+}
