+++ source/scripts/Background/utils/startPendingTransactionsPolling.ts
@@ -1,0 +1,5 @@
+export function startPendingTransactionsPolling() {
+  chrome.alarms.create('check_pending_transactions', {
+    periodInMinutes: 120, //run after 2 hours
+  });
+}
